requirements.txt
mpi4py
Pillow
numpy

--extra-index-url https://download.pytorch.org/whl/cu121

torch
torchvision
torchaudio
